Add ASCII word wrap function.
currently only supports splitting on spaces, it searches for a locally
optimal line break and therefore will result in lines at the end that
could be much shorter than earlier lines.

there is no emergency mode for wrapping an otherwise unwrappable string,
for example a long string with no spaces in it.