Add basic cycle detection
uses an ArrayDeque as a stack. currently this is not thread safe and
will need to be modified when a concurrent strategy is implemented

Adds and removes the `System.identityHashCode()` from a stack. when the
hash code is detected in the list of currently processing nodes, a cycle
has been detected. Add the hash code to a list of identifiers (in case
we encounter many cycles in a single node tree, and then immediately
return.

This is faster than object comparison and has relatively low memory
overhead, but is most likely not optimal